+++ lib/flutter_geofire.dart
@@ -46,6 +46,18 @@ class Geofire {
     return isSet;
   }
 
+  static Future<bool?> removeGeoQueryEventListener() async {
+    final bool? isSet = await _channel
+        .invokeMethod('removeGeoQueryEventListener', <String, dynamic>{});
+    return isSet;
+  }
+
+  static Future<bool?> removeGeoQueryDataEventListener() async {
+    final bool? isSet = await _channel
+        .invokeMethod('removeGeoQueryDataEventListener', <String, dynamic>{});
+    return isSet;
+  }
+
   static Future<Map<String, dynamic>> getLocation(String id) async {
     final Map<dynamic, dynamic> response = await (_channel
         .invokeMethod('getLocation', <String, dynamic>{"id": id}));